@@ -4,7 +4,10 @@
     {
         [EnumName(FromTranslation = "Common.Automatic")]
         Default,
-        D3D11,
         D3D10,
+        D3D11,
+        OpenGL,
+        Vulkan,
+        Metal
     }
 }
