@@ -363,12 +363,6 @@ namespace Bloxstrap
             if (!File.Exists(Paths.Application) || Paths.Process == Paths.Application)
                 return;
 
-            // 2.0.0 downloads updates to <BaseFolder>/Updates so lol
-            bool isAutoUpgrade = App.LaunchSettings.UpgradeFlag.Active
-                || Paths.Process.StartsWith(Path.Combine(Paths.Base, "Updates"))
-                || Paths.Process.StartsWith(Path.Combine(Paths.LocalAppData, "Temp"))
-                || Paths.Process.StartsWith(Paths.TempUpdates);
-
             var existingVer = FileVersionInfo.GetVersionInfo(Paths.Application).ProductVersion;
             var currentVer = FileVersionInfo.GetVersionInfo(Paths.Process).ProductVersion;
 
@@ -377,28 +371,24 @@ namespace Bloxstrap
 
             if (currentVer is not null && existingVer is not null && Utilities.CompareVersions(currentVer, existingVer) == VersionComparison.LessThan)
             {
-                var result = Frontend.ShowMessageBox(
+                var resultVer = Frontend.ShowMessageBox(
                     Strings.InstallChecker_VersionLessThanInstalled,
                     MessageBoxImage.Question,
                     MessageBoxButton.YesNo
                 );
 
-                if (result != MessageBoxResult.Yes)
+                if (resultVer != MessageBoxResult.Yes)
                     return;
             }
 
-            // silently upgrade version if the command line flag is set or if we're launching from an auto update
-            if (!isAutoUpgrade)
-            {
-                var result = Frontend.ShowMessageBox(
-                    Strings.InstallChecker_VersionDifferentThanInstalled,
-                    MessageBoxImage.Question,
-                    MessageBoxButton.YesNo
-                );
+            var result = Frontend.ShowMessageBox(
+                Strings.InstallChecker_VersionDifferentThanInstalled,
+                MessageBoxImage.Question,
+                MessageBoxButton.YesNo
+            );
 
-                if (result != MessageBoxResult.Yes)
-                    return;
-            }
+            if (result != MessageBoxResult.Yes)
+                return;
 
             App.Logger.WriteLine(LOG_IDENT, "Doing upgrade");
 
@@ -506,27 +496,13 @@ namespace Bloxstrap
                     if (App.Settings.Prop.BootstrapperStyle == BootstrapperStyle.ClassicFluentDialog)
                         App.Settings.Prop.BootstrapperStyle = BootstrapperStyle.FluentDialog;
 
-                    _ = int.TryParse(App.FastFlags.GetPreset("Rendering.Framerate"), out int x);
+                    _ = int.TryParse(App.FastFlags.GetPreset("Rendering.Framerate1"), out int x);
                     if (x == 0)
-                        App.FastFlags.SetPreset("Rendering.Framerate", null);
+                        App.FastFlags.SetPreset("Rendering.Framerate1", null);
                 }
 
                 if (Utilities.CompareVersions(existingVer, "2.8.0") == VersionComparison.LessThan)
                 {
-                    if (isAutoUpgrade)
-                    {
-                        if (App.LaunchSettings.Args.Length == 0)
-                            App.LaunchSettings.RobloxLaunchMode = LaunchMode.Player;
-
-                        string? query = App.LaunchSettings.Args.FirstOrDefault(x => x.Contains("roblox"));
-
-                        if (query is not null)
-                        {
-                            App.LaunchSettings.RobloxLaunchMode = LaunchMode.Player;
-                            App.LaunchSettings.RobloxLaunchArgs = query;
-                        }
-                    }
-
                     string oldDesktopPath = Path.Combine(Paths.Desktop, "Play Roblox.lnk");
                     string oldStartPath = Path.Combine(Paths.WindowsStartMenu, "Bloxstrap");
 
@@ -611,21 +587,11 @@ namespace Bloxstrap
 
             App.SendStat("installAction", "upgrade");
 
-            if (isAutoUpgrade)
-            {
-#pragma warning disable CS0162 // Unreachable code detected
-                if (OpenReleaseNotes)
-                    Utilities.ShellExecute($"https://github.com/{App.ProjectRepository}/wiki/Release-notes-for-Bloxstrap-v{currentVer}");
-#pragma warning restore CS0162 // Unreachable code detected
-            }
-            else
-            {
-                Frontend.ShowMessageBox(
-                    string.Format(Strings.InstallChecker_Updated, currentVer),
-                    MessageBoxImage.Information,
-                    MessageBoxButton.OK
-                );
-            }
+            Frontend.ShowMessageBox(
+                string.Format(Strings.InstallChecker_Updated, currentVer),
+                MessageBoxImage.Information,
+                MessageBoxButton.OK
+            );
         }
     }
 }
