@@ -4,14 +4,13 @@ namespace Bloxstrap.UI.ViewModels.About
 {
     public class AboutViewModel : NotifyPropertyChangedViewModel
     {
-        public string Version => string.Format(Strings.Menu_About_Version, App.Version);
+        public string Version => string.Format(Strings.Menu_About_Version, App.ShortCommitHash);
 
         public BuildMetadataAttribute BuildMetadata => App.BuildMetadata;
 
         public string BuildTimestamp => BuildMetadata.Timestamp.ToFriendlyString();
-        public string BuildCommitHashUrl => $"https://github.com/{App.ProjectRepository}/commit/{BuildMetadata.CommitHash}";
+        public string BuildCommitHashUrl => $"{App.ProjectRepository}/commit/{BuildMetadata.CommitHash}";
 
-        public Visibility BuildInformationVisibility => App.IsProductionBuild ? Visibility.Collapsed : Visibility.Visible;
         public Visibility BuildCommitVisibility => App.IsActionBuild ? Visibility.Visible : Visibility.Collapsed;
     }
 }
