@@ -1,6 +1,6 @@
 ﻿namespace Bloxstrap.UI.ViewModels.Settings
 {
-    public class BehaviourViewModel : NotifyPropertyChangedViewModel
+    public class BootstrapperViewModel : NotifyPropertyChangedViewModel
     {
         public bool ConfirmLaunches
         {
@@ -12,6 +12,12 @@
         {
             get => App.Settings.Prop.ForceRobloxLanguage;
             set => App.Settings.Prop.ForceRobloxLanguage = value;
+        }
+
+        public bool WPFSoftwareRenderEnabled
+        {
+            get => App.Settings.Prop.WPFSoftwareRender;
+            set => App.Settings.Prop.WPFSoftwareRender = value;
         }
 
         public bool BackgroundUpdates
@@ -26,6 +32,12 @@
         {
             get => App.State.Prop.ForceReinstall || IsRobloxInstallationMissing;
             set => App.State.Prop.ForceReinstall = value;
+        }
+
+        public string SelectedChannel
+        {
+            get => App.Settings.Prop.CustomChannel;
+            set => App.Settings.Prop.CustomChannel = value;
         }
     }
 }
